+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Config options for tonic boost amounts
+
 ## [1.21.1-0.2.4] - 2025-03-09
 ### Changed
 - EvalEx library is now shadowed to a new package, which may help prevent mod conflicts